+++ src/scenes/week/week.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import Video from './_/video';
-import Layout from '../../shared/week_layout';
+import Layout from '../../shared/layout';
 import $ from "jquery";
 import { Grid, Nav, NavItem, Glyphicon } from 'react-bootstrap';
 import FacebookProvider, { Comments } from 'react-facebook';
@@ -33,6 +33,12 @@ class WorkShop extends Component {
     return (
       <div className="week">
         <Layout>
+          <header>
+            <Grid>
+              <h1>I SEMANA DO INFLUENCIADOR DIGITAL</h1>
+              <h2>Transformando pessoas comuns em Influenciadores Digitais</h2>
+            </Grid>
+          </header>
           <Grid>
             <div className="navbar-header">
               <button data-toggle="collapse-side" data-target=".side-collapse" data-target-2=".side-collapse-container" type="button" className="navbar-toggle pull-left"><span className="icon-bar"></span><span className="icon-bar"></span><span className="icon-bar"></span></button>
