+++ src/shared/track/index.js
@@ -1,0 +1,3 @@
+//index.js
+import Root from './tracker'
+export default Root
