--- src/shared/week_layout/_/footer.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import Footer from '../../layout/_/footer';
-
-export default () => {
-  return (
-    <Footer />
-  );
-}
