--- src/shared/week_layout/_/header.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import Header from '../../layout/_/header';
-import { Grid } from 'react-bootstrap';
-
-export default () => {
-  return (
-    <div>
-      <Header />
-      <header>
-        <Grid>
-          <h1>I SEMANA DO INFLUENCIADOR DIGITAL</h1>
-          <h2>Transformando pessoas comuns em Influenciadores Digitais</h2>
-        </Grid>
-      </header>
-    </div>
-  );
-}
