--- src/shared/week_layout/index.js
@@ -1,3 +1,0 @@
-//index.js
-import Root from './week_layout'
-export default Root
