--- src/shared/week_layout/week_layout.css
@@ -1,37 +1,0 @@
-h1 {
-  color: white;
-  padding-top: 41px;
-  font-weight: bold;
-}
-
-h2 {
-  color: white;
-  margin-top: -13px;
-  margin-bottom: 70px;
-  font-size: 25px;
-}
-
-header {
-  font-family: 'Open Sans', sans-serif;
-  text-align: center;
-  background-image: url(https://firebasestorage.googleapis.com/v0/b/***REMOVED***.appspot.com/o/bg-workshop.jpg?alt=media&token=859bc49f-7d49-4b07-a52f-c20dd5bdb5cd);
-  background-attachment: scroll;
-  background-position: center;
-  background-size: cover;
-  padding: 0;
-}
-
-.menu {
-  margin-top: 40px;
-}
-
-@media screen and (max-width: 768px) {
-  h2 {
-    font-size: 15px;
-    margin-top: -10px;
-  }
-
-  h1 {
-    font-size: 17px;
-  }
-}
